change timestamp from int to float

+++ code/python-api-server/app.py
@@ -219,9 +219,9 @@ def stream(task_id):
         last_sent = 0
         while True:
             progress = get_task_progress(task_id)
-            if int(progress["timestamp"]) > last_sent:
+            if float(progress["timestamp"]) > last_sent:
                 yield f"data: {progress}\n\n"
-                last_sent = int(progress["timestamp"])
+                last_sent = float(progress["timestamp"])
             if progress["status"] in ("SUCCESS", "FAILURE"):
                 break
             time.sleep(1)
